+++ lib/doorkeeper/config.rb
@@ -437,6 +437,24 @@ module Doorkeeper
              end
            end)
 
+    # Allows the client to not pass a redirect_uri parameter to the
+    # authorization endpoint during the authorization code flow and implicit
+    # flow. If the client does not pass a redirect_uri parameter to the
+    # authorization endpoint then then Doorkeeper will redirect the resource
+    # owner to the redirection endpoint specified when the application was
+    # created, as per Section 3.1.2 of the OAuth 2.0 specification:
+    # https://datatracker.ietf.org/doc/html/rfc6749#section-3.1.2
+    #
+    # By default this option is false, meaning that the client must pass a
+    # redirect_uri parameter to the authorization endpoint.
+    #
+    # This option should only be set to true if your application does not have
+    # a NOT_NULL constraint applied to the redirect_uri column of its
+    # oauth_access_grants database table.
+    #
+    option :redirect_uri_optional_during_authorization,
+           default: false
+
     attr_reader :reuse_access_token,
                 :token_secret_fallback_strategy,
                 :application_secret_fallback_strategy