+++ spec/dummy/db/schema.rb
@@ -18,7 +18,7 @@ ActiveRecord::Schema.define(version: 20180210183654) do
     t.integer "application_id", null: false
     t.string "token", null: false
     t.integer "expires_in", null: false
-    t.text "redirect_uri", null: false
+    t.text "redirect_uri"
     t.datetime "created_at", null: false
     t.datetime "revoked_at"
     t.string "scopes"